+++ main.go
@@ -9,7 +9,7 @@ import (
 	"github.com/mlange-42/xwrd/util"
 )
 
-const version = "0.1.0-dev"
+const version = "0.1.1-dev"
 
 func main() {
 	util.EnsureDirs()