+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
-## [[unpublished]](https://github.com/mlange-42/xwrd/compare/...main)
+## [[unpublished]](https://github.com/mlange-42/xwrd/compare/v0.1.0...main)
 
-Nothing
+### Features
+
+* Filter anagrams by a pattern (#9)
+* Interactively change anagram mode and flags (#9)
+
+## [[v0.1.0]]